Added new pkg and class AppWideExceptionHandler for handle  errors and refactoured customer impl.

+++ SpringFramework/15_Spring_Data_JPA/src/main/java/lk/ijse/spring/service/impl/CustomerServiceImpl.java
@@ -31,11 +31,17 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void addCustomer(CustomerDTO cus) {
+        if (repo.existsById(cus.getId())) {
+            throw new RuntimeException("Customer "+cus.getId()+" Already Exist..!");
+        }
         repo.save(mapper.map(cus, Customer.class));
     }
 
     @Override
     public void deleteCustomer(String id) {
+        if (!repo.existsById(id)) {
+            throw new RuntimeException("Customer "+id+" Not Available to Delete !!!.");
+        }
         repo.deleteById(id);
     }
